User Api v0.2

+++ src/main/java/com/backend/backendfinalproject/repositories/UserRepositoryImpl.java
@@ -1,6 +1,8 @@
 package com.backend.backendfinalproject.repositories;
 
+import com.backend.backendfinalproject.models.Response;
 import com.backend.backendfinalproject.models.User;
+import com.backend.backendfinalproject.repositories.interfaces.IUserRepository;
 import de.mkammerer.argon2.Argon2;
 import de.mkammerer.argon2.Argon2Factory;
 import org.hibernate.Session;
@@ -8,6 +10,7 @@ import org.springframework.stereotype.Repository;
 
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
+import javax.persistence.Query;
 import javax.transaction.Transactional;
 import java.util.List;
 
@@ -27,28 +30,88 @@ public class UserRepositoryImpl implements IUserRepository {
     }
 
     @Override
-    public void remove(int id) {
+    public Object getUser(int id) {
+        Session session = entityManager.unwrap(Session.class);
+        List<User> usersList = session.createQuery("FROM User WHERE id = :id", User.class).setParameter("id", id).getResultList();
+        return !usersList.isEmpty() ? usersList.get(0) : new Response("User not found", false);
+    }
+
+    @Override
+    public Response remove(int id) {
         User user = entityManager.find(User.class, id);
-        entityManager.remove(user);
+        Response response = new Response();
+        response.setStatus(false);
+
+        if (user != null) {
+            response.setStatus(true);
+            response.setMessage("Complete deletion");
+            entityManager.remove(user);
+        } else {
+            response.setMessage("Failed deletion");
+        }
+
+        return response;
     }
 
     @Override
-    public void register(User user) {
-        entityManager.merge(user);
+    public Response register(User user) {
+        User newUser = entityManager.merge(user);
+        Response rsp = new Response();
+        rsp.setStatus(false);
+
+        if (newUser != null) {
+            rsp.setMessage("Registration Completed");
+            rsp.setStatus(true);
+        } else {
+            rsp.setMessage("Registration Failed");
+        }
+
+        return rsp;
     }
 
     @Override
-    public boolean validateCredentials(User user) {
+    public User getUserByCredentials(User user) {
         Session session = entityManager.unwrap(Session.class);
         List<User> listUsers = session.createQuery("FROM User WHERE email = :email", User.class)
                 .setParameter("email", user.getEmail())
                 .getResultList();
 
-        if (listUsers.isEmpty()) return false;
+        if (listUsers.isEmpty()) return null;
 
         String passwordHashed = listUsers.get(0).getPassword();
         Argon2 argon2 = Argon2Factory.create(Argon2Factory.Argon2Types.ARGON2id);
 
-        return argon2.verify(passwordHashed, user.getPassword());
+        return argon2.verify(passwordHashed, user.getPassword()) ? listUsers.get(0) : null;
+    }
+
+    @Override
+    public Response update(User user) {
+        Session session = entityManager.unwrap(Session.class);
+        Response response = new Response();
+        response.setStatus(false);
+
+        Query query = session.createQuery("UPDATE User SET avatar= :avatar, direction = :direction, email = :email, last_name = :last_name, name = :name, password = :pass, phone = :phone, sex = :sex, state = :state, city = :city WHERE id = :id");
+        query.setParameter("avatar", user.getAvatar());
+        query.setParameter("direction", user.getDirection());
+        query.setParameter("email", user.getEmail());
+        query.setParameter("last_name", user.getLast_name());
+        query.setParameter("name", user.getName());
+        query.setParameter("pass", user.getPassword());
+        query.setParameter("phone", user.getPhone());
+        query.setParameter("sex", user.getSex());
+        query.setParameter("state", user.getState());
+        query.setParameter("city", user.getCity());
+        query.setParameter("id", user.getId());
+
+        int status = query.executeUpdate();
+
+        if (status == 1) {
+            response.setMessage("Update Success");
+            response.setStatus(true);
+        } else {
+            response.setMessage("Update Failed");
+        }
+
+        return response;
     }
 }
